+++ FolderDiffIL4DotNet.Tests/Runner/CliOverrideApplierTests.cs
@@ -124,7 +124,9 @@ namespace FolderDiffIL4DotNet.Tests.Runner
             Assert.Contains("buildserver1_", builder.ILIgnoreLineContainingStrings);
             Assert.Contains("// Code size ", builder.ILIgnoreLineContainingStrings);
             Assert.Contains("¥temp¥develop¥", builder.ILIgnoreLineContainingStrings);
+            Assert.Contains("¥¥temp¥¥develop¥¥", builder.ILIgnoreLineContainingStrings);
             Assert.Contains(@"\temp\develop\", builder.ILIgnoreLineContainingStrings);
+            Assert.Contains(@"\\temp\\develop\\", builder.ILIgnoreLineContainingStrings);
             Assert.Contains("/temp/develop/", builder.ILIgnoreLineContainingStrings);
         }
 
@@ -138,6 +140,8 @@ namespace FolderDiffIL4DotNet.Tests.Runner
 
             Assert.True(builder.ShouldIgnoreILLinesContainingConfiguredStrings);
             Assert.Contains("buildserver1_", builder.ILIgnoreLineContainingStrings);
+            Assert.Contains("¥¥temp¥¥develop¥¥", builder.ILIgnoreLineContainingStrings);
+            Assert.Contains(@"\\temp\\develop\\", builder.ILIgnoreLineContainingStrings);
             Assert.Contains("/temp/develop/", builder.ILIgnoreLineContainingStrings);
         }